% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{add_bearings}
\alias{add_bearings}
\alias{add_distances}
\title{Add estimated bearings or distances data to a capthist object}
\usage{
add_bearings(capthist, value)

add_distances(capthist, value)
}
\arguments{
\item{capthist}{a \code{\link{capthist}} object}

\item{value}{estimated bearings or distances data}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

