% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-import.R
\name{add_covariates}
\alias{add_covariates}
\title{Add spatial covariates to mask or traps}
\usage{
add_covariates(x, data)
}
\arguments{
\item{x}{a \link[gibbonsecr]{gmask} or
\link[gibbonsecr]{gcapthist} object.}

\item{data}{a \link[gibbonsecr]{gshp} object}
}
\description{
Based on the \link[secr]{addCovariates} function from \pkg{secr}
  but which also accommodates \link[sp]{SpatialPointsDataFrame} data objects.
}
\examples{
\dontrun{

library(secr)
library(ggplot2)
data(N.annamensis)

mask = make_mask(N.annamensis.capthist, spacing = 250, buffer = 6000)
head(covariates(mask[[1]]))

ggplot() + coord_fixed() +
    geom_mask(mask) +
    geom_capthist(N.annamensis.capthist) +
    labs(x = "Longitude", y = "Latitude")

mask = add_covariates(mask, N.annamensis.habitat)
head(covariates(mask[[1]]))

ggplot() + coord_fixed() +
    geom_mask(mask, covariate = "habitat") +
    geom_capthist(N.annamensis.capthist) +
    scale_fill_manual(values = c("darkgreen", "lightgreen")) +
    labs(x = "Longitude", y = "Latitude")
}


}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

