% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-classes.R
\name{gibbonsecr-classes}
\alias{gcapthist}
\alias{gibbonsecr-classes}
\alias{gmask}
\alias{gsecr}
\alias{gshp}
\title{gibbonsecr object classes}
\description{
Object classes returned from \pkg{gibbonsecr} functions.
}
\section{gcapthist}{

  Object containing recapture data plus estimated bearings (if present) and
  estimated distances (if present). Returned from from
  \link[gibbonsecr]{import_data} and based on the \link[secr]{capthist} class
  from the \pkg{\link[secr]{secr}} package. Objects of this type are always
  multi-session and each capthist element has \code{bearings} and/or
  \code{distances} attributes which are arrays of equal dimension to the
  recapture data but contain estimated bearings and/or estimated distances
  (if these data are not present then these attributes are set to
  \code{NULL}).
}

\section{gmask}{

  Returned from \link[gibbonsecr]{make_mask} and based on the
  \link[secr]{mask} class from the \pkg{\link[secr]{secr}} package. Objects of this
  type are always multi-session \code{"trapbuffer"} masks and have
  \code{buffer} and \code{spacing} attributes.
}

\section{gshp}{

  Returned from \link[gibbonsecr]{import_shp}. Objects of this type are lists
  with two elements: \code{sp} which contains an object of class
  \link[sp]{sp}, \code{fortify} which contains an object returned from the
  \link[ggplot2]{fortify.sp} function from the \pkg{ggplot2} package, applied
  to the sp object (this is used by the \link[gibbonsecr]{geom_shp}
  function).
}

\section{gsecr}{

  Returned from \link[gibbonsecr]{gfit} and contain information on the model
  fit such as paramter estimates.
}

