% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{delta_method}
\alias{delta_method}
\title{Delta method}
\usage{
delta_method(f, beta, vcov, ...)
}
\arguments{
\item{f}{Function of vector beta}

\item{beta}{Parameter vector}

\item{vcov}{Variance-covariance matrix for beta}

\item{...}{additional arguments to pass to \code{f}}
}
\description{
Function to estimate the variance of nonlinear functions of normal variables
}
\details{
Uses the \code{stats::numericDeriv} function.
}
\examples{
\dontrun{
# Example:
hn = function(beta, r){
   g0 <- invlogit(beta[1])
   sigma <- exp(beta[2])
   g0 * exp(-r^2/2/sigma^2)
}
beta = c(logit(0.5), log(1000))
vcov = matrix(c(4.337749e-02,-1.869121e-03,-1.869121e-03,7.368978e-04), 2, 2) ; vcov
r = seq(0, 3000, length = 25)
delta = delta_method(f, beta, vcov, r = r)
plot(r, delta$est, type = "l", xlim = c(0,3000), ylim = c(0,1))
lines(r, delta$est + 1.96 * delta$se, lty = 2, col = 2, lwd = 2)
lines(r, delta$est - 1.96 * delta$se, lty = 2, col = 2, lwd = 2)
}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

