% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{delta_method_Xbeta}
\alias{delta_method_Xbeta}
\title{Delta method for linear models}
\usage{
delta_method_Xbeta(X, beta, vcov)
}
\arguments{
\item{X}{Design matrix}

\item{beta}{Parameter vector}

\item{vcov}{Variance-covariance matrix for beta}
}
\description{
Function to estimate confidence intervals for functions of estimated parameters
}
\details{
This is a wrapper function for \code{delta_method}
}
\examples{
\dontrun{
# Examples:
n = 100
x = seq(0, 1, length = n)
y = x^2 + rnorm(n, 0, 0.1)
plot(x, y)
fit = lm(y ~ I(x^2))

# conventional approach
preds = predict(fit, se = TRUE)
lines(x, preds$fit, col = 4, lwd = 2)
lines(x, preds$fit + qnorm(0.05/2) * preds$se, lty = 2, col = 4, lwd = 2)
lines(x, preds$fit + qnorm(1-0.05/2) * preds$se, lty = 2, col = 4, lwd = 2)

# delta method
X = model.matrix(fit)
beta = coef(fit)
vcov = vcov(fit)
intervals = delta_method_Xbeta(X, beta, vcov)
lines(x, intervals$est, col = 2, lwd = 2)
lines(x, intervals$lower, lty = 2, col = 2, lwd = 2)
lines(x, intervals$upper, lty = 2, col = 2, lwd = 2)
}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

