% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{get_bearings}
\alias{get_bearings}
\alias{get_captures}
\alias{get_distances}
\title{Extract data from capthist}
\usage{
get_bearings(capthist)

get_distances(capthist)

get_captures(capthist, summarise = NULL)
}
\arguments{
\item{capthist}{a \code{\link{capthist}} object}

\item{summarise}{choose whether to collapse capture data across traps or occasions (see \bold{Details})}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

