% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-package.r
\docType{package}
\name{gibbonsecr-package}
\alias{gibbonsecr-package}
\alias{gibbonsecr}
\title{SECR for acoustic gibbon surveys}
\description{
An implementation of Spatially Explicit Capture-Recapture (SECR) for
estimating the density of gibbon groups from acoustic surveys.

This package also contains a user interface which can be accessed using the
\code{\link[gibbonsecr]{gui}} function. See the online manual for more details
\url{http://dkidney.github.io/gibbonsecr}.

\pkg{gibbonsecr} relies on several functions from the \pkg{\link[secr]{secr}} package
and uses an analagous system of objects. Model fitting is performed in a similar way
using a fine grid of points called a mask to approximate the spatial integration
component of the SECR likelihood. The defintion of 'trap' and 'session' are also
consistent with \pkg{secr}. However \pkg{gibbonsecr} differs from \pkg{secr} in the
following respects:
\enumerate{
\item Supplementary information on detected animals in the form of estimated
bearings and estimated distances can be incorporated in addition to the recapture
data to improve precision in the density estimate.
\item Stochastic availability of animals can be modelled using the parameter
\code{pcall}. In general, this represents the proportion of animals being availablity
for detection on each survey occasion. For gibbons, it can be interpreted as the
proportion of groups that call on a given survey day.
\item Trap data objects associated with capture history objects are always
of \code{type} \code{\link[secr]{proximity}}.
\item \strong{no group cov}.
\item \strong{session cov and timecov attributes}.
}
The main functions in \pkg{gibbonsecr} are:
\tabular{ll}{
\code{\link[gibbonsecr]{import_data}} \tab importing survey data, listening post
locations and non-spatial covariates \cr
\code{\link[gibbonsecr]{import_shp}}  \tab importing GIS shapefiles (\code{.shp}) \cr
\code{\link[gibbonsecr]{make_mask}}   \tab constructing mask objects \cr
\code{\link[gibbonsecr]{gfit}}        \tab fitting models \cr
}
See \code{\link[gibbonsecr]{N.annamensis}} for an example data set from a survey of
\emph{Nomascus annamensis} in northwestern Cambodia.
}
\examples{
\dontrun{

library(gibbonsecr)
par(mfrow = c(1,1), mar = c(4,4,2,1))


# import survey data -----

folder = system.file("extdata/N.annamensis", package = "gibbonsecr")
capthist = import_data(
    detections = file.path(folder, "detections.csv"),
    posts      = file.path(folder, "posts.csv"),
    covariates = file.path(folder, "covariates.csv"),
    details    = list(bearings = list(units = "degrees", type = "continuous"))
)
summary(capthist)
plot(capthist)


# import shapefiles -----

region = import_shp(file.path(folder, "region.shp"))
plot(region)
points(capthist)

habitat = import_shp(file.path(folder, "habitat.shp"))
plot(habitat)
plot(region, add = TRUE)
points(capthist)


# make mask -----

mask1 = make_mask(capthist, buffer = 6000, spacing = 100)
summary(mask1)
plot(mask1)
points(capthist)

mask2 = make_mask(capthist, buffer = 6000, spacing = 100, poly = region$sp)
summary(mask2)
plot(mask2)
points(capthist)

mask3 = add_covariates(mask2, habitat)
summary(mask3)
plot(mask3, covariate = "habitat", col = c("darkgreen","lightgreen"))
points(capthist)


# fit model ----

fit1 = gfit(capthist = capthist, mask = mask1)
summary(fit1)

fit2 = gfit(capthist = capthist, mask = mask2)
summary(fit2)

fit3 = gfit(capthist = capthist, mask = mask3, model = list(D ~ habitat))
summary(fit3)

AIC(fit2)
AIC(fit3)


# plot results -----

plot(fit3)
plot(fit3, ci = TRUE)
plot(fit3, ci = TRUE, method = "boot")

plot(fit3, "bearings")
plot(fit3, "bearings", ci = TRUE)
plot(fit3, "bearings", ci = TRUE, method = "boot")

plot(fit3, "detsurf", session = 7)
points(fit3$capthist, session = 7)

plot(fit3, "densurf")
points(fit3$capthist)


# get predictions ----

newdata = data.frame(habitat = c("primary", "secondary"))
predict(fit3, submodels = "D", newdata = newdata)
}
}
\references{
Kidney D., Rawson B.M., Borchers D.L., Stevenson B.C., Marques T.A., & Thomas L.
2016. An Efficient Acoustic Density Estimation Method with Human Detectors Applied
to Gibbons in Cambodia. \emph{PLoS ONE} 11(5): e0155066.

Borchers, D. L., Stevenson, B. C., Kidney, D., Thomas, L., & Marques, T. A. 2015. A
Unifying Model for Capture-Recapture and Distance Sampling Surveys of Wildlife
Populations. \emph{Journal of the American Statistical Association}, 110(509),
195-204.
}
\seealso{
\link[gibbonsecr]{N.annamensis}, \link[gibbonsecr]{import_data},
\link[gibbonsecr]{gfit}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
