% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{gibbonsecr_fit}
\alias{gibbonsecr_fit}
\title{Fit an SECR model to acoustic gibbon survey data}
\usage{
gibbonsecr_fit(capthist, model = list(), mask = NULL, fixed = list(),
  model.options = list(), mask.options = list(), fitting.options = list(),
  start = NULL, trace = FALSE, debug = TRUE)
}
\arguments{
\item{capthist}{a \code{\link{capthist}} object}

\item{model}{TODO}

\item{mask}{TODO}

\item{fixed}{TODO}

\item{model.options}{a list of control options determining the type of model to be fitted (see Details).}

\item{mask.options}{TODO}

\item{fitting.options}{TODO}

\item{start}{TODO}

\item{trace}{TODO}

\item{debug}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
## Not run: 

data(N.annamensis)

fit = gibbonsecr_fit(N.annamensis)

AIC(fit)
coef(fit)
#confint(fit)
#plot(fit)
#predict(fit)
#print(fit)
vcov(fit)

## End(Not run)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

