% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-gui.r
\name{gibbonsecr_gui}
\alias{gibbonsecr_gui}
\title{Launch the graphical user interface}
\usage{
gibbonsecr_gui(prompt.save.on.exit = FALSE, quit.r.on.exit = FALSE)
}
\arguments{
\item{prompt.save.on.exit}{logical scalar, determining whether a save workspace prompt message should be shown when the GUI window is closed (defaults to \code{FALSE})}

\item{quit.r.on.exit}{a logical scalar, determining whether the background R process should quit when the GUI window is closed (defaults to \code{FALSE})}
}
\description{
Graphical user interface for fitting acoustic gibbon survey data using
  the gibbonsSECR package
}
\details{
Currently allows the inclusion of the following components:
  \itemize{

  \item{half normal and hazard rate detection function (with or without g0)}

  \item{von Mises and wrapped Cauchy distributions for estimated bearings}

  \item{gamma and log-normal distributions for estimated distances}

  \item{calling probability for multi-occasion surveys (i.e. surveys with at
  least two consecutive days of data at each set of listening posts, with
  recaptures made across days)}

  \item{covariates for all model parameters except the hazard rate shape
  parameter (z)}

  \item{saving and reloading of workspace}

  }
}
\section{Things still to be added}{


  \itemize{

  \item{binned distance estimates}

  \item{survey region and habitat polygons}

  \item{spline-based models for the density surface}

  \item{manual}

  }
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

