% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{import_data}
\alias{import_data}
\title{Import survey data from csv files}
\usage{
import_data(detections, posts, covariates = NULL, details = list())
}
\arguments{
\item{detections}{file containing the recapture data and the estiamated
bearings and distances}

\item{posts}{file containing the locations of the posts}

\item{covariates}{optional file containing the covariate data}

\item{details}{optional argument determining the type and units of the
estiamted bearings and distances}
}
\description{
TODO
}
\details{
Details - file paths will have been obtained via the browse button
  in the GUI
}
\examples{
library(sp)
library(secr)
wd = setwd(system.file("extdata/N.annamensis", package = "gibbonsecr"))
op = par(no.readonly = TRUE)

capthist = import_data(
    detections = "example_detections_file.csv",
    posts      = "example_posts_file.csv",
    covariates = "example_covariates_file.csv"
)
class(capthist)

region = import_shp("region")

par(mfrow = c(1,1), mar = c(0,0,0,0))
plot(region)
plot(traps(capthist), add = TRUE)

setwd(wd)
par(op)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

