% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{import_shp}
\alias{import_shp}
\title{Import region and habitat shapefiles}
\usage{
import_shp(filepath, type = c("region", "habitat"), verbose = FALSE, ...)
}
\arguments{
\item{filepath}{TODO}

\item{type}{TODO}

\item{verbose}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

