% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mask.r
\name{make_mask}
\alias{make_mask}
\title{Make a habitat mask}
\usage{
make_mask(traps, buffer = 6000, spacing = 250, regular = TRUE,
  poly = NULL)
}
\arguments{
\item{traps}{a \link[secr]{traps}, \link[secr]{capthist} or
\link[gibbonsecr]{gcapthist} object (traps must be \link[secr]{proximity}
type)}

\item{buffer}{width of buffer (in metres)}

\item{spacing}{spacing between grid points (metres)}

\item{regular}{if \code{TRUE}, then all mask points will be taken from a
single regular grid}

\item{poly}{bounding polygon to which mask should be clipped (see
documentation for \link[secr]{make.mask})}
}
\description{
A wrapper function for \link[secr]{make.mask} (from the
\pkg{\link[secr]{secr}} package) which returns a
\link[gibbonsecr]{gmask} object.
}
\seealso{
\link[gibbonsecr]{add_covariates}, \link[gibbonsecr]{plot.gmask}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
