% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-N.annamensis.r
\docType{data}
\name{N.annamensis}
\alias{N.annamensis}
\alias{N.annamensis.capthist}
\alias{N.annamensis.fit}
\alias{N.annamensis.habitat}
\alias{N.annamensis.mask}
\alias{N.annamensis.region}
\title{Example survey data}
\source{
Survey data were provided by Dr Ben Rawson, IUCN SSC Primate Specialist Group,
  Southeast Asia/Indochina.
}
\usage{
data(N.annamensis)
}
\description{
Data from 2010 survey of \emph{Nomascus annamensis}, Veun Sai Reserve, NW
  Cambodia.
}
\details{
The following objects are provided:
\tabular{ll}{
\code{N.annamensis.region} \tab an imported GIS polygon for the survey region (class
\link[gibbonsecr]{gshp}) \cr
\code{N.annamensis.habitat} \tab artificial polygon containing habitat data (class
\link[gibbonsecr]{gshp}) \cr
\code{N.annamensis.capthist} \tab survey data -- detections and estimated bearings
(class \link[gibbonsecr]{gcapthist}) \cr
\code{N.annamensis.mask} \tab a mask object (class \link[gibbonsecr]{gmask}) \cr
\code{N.annamensis.fit} \tab a fitted model (class \link[gibbonsecr]{gsecr}) \cr
}
}
\references{
Thinh, V.N., Mootnick, A.R., Thanh, V.N., & Roos, C. 2010. A new species of crested
  gibbon, from the central Annamite mountain range. \emph{Vietnamese Journal of
  Primatology}, 1(4):112.

  Kidney D., Rawson B.M., Borchers D.L., Stevenson B.C., Marques T.A., & Thomas L.
  2016. An Efficient Acoustic Density Estimation Method with Human Detectors Applied to
  Gibbons in Cambodia. \emph{PLoS ONE} 11(5): e0155066.
}

