% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-data.r
\docType{data}
\name{N.siki}
\alias{N.siki}
\title{TODO}
\format{A object of class \code{\link{capthist}}.}
\source{
TODO
}
\usage{
data(N.annamensis)
}
\description{
TODO
}
\references{
TODO
}

