% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot.gibbonsecr_fit}
\alias{plot.gibbonsecr_fit}
\title{Plot a fitted gibbon secr model}
\usage{
\method{plot}{gibbonsecr_fit}(fit, which = c("detectfn", "pdot", "bearings",
  "distances", "density"), session = 1, use.global.par.settings = FALSE,
  ...)
}
\arguments{
\item{which}{TODO}

\item{session}{TODO}

\item{use.global.par.settings}{TODO}

\item{...}{additional arguments (not used)}

\item{CI}{TODO}

\item{nx}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

