% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot.gibbonsecr_sim}
\alias{plot.gibbonsecr_sim}
\title{Plot simulation results}
\usage{
\method{plot}{gibbonsecr_sim}(x, ..., CI = FALSE, exp = FALSE,
  use.global.par.settings = FALSE)
}
\arguments{
\item{x}{a \code{gibbonsecr} model object (i.e. an object returned from the \code{\link{gibbonsecr_fit}} function)}

\item{...}{additional arguments (not used)}

\item{CI}{TODO}

\item{exp}{TODO}

\item{use.global.par.settings}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

