% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.r
\name{plot.gcapthist}
\alias{geom_capthist}
\alias{plot.gcapthist}
\alias{points.gcapthist}
\title{Plot a gibbonsecr capthist}
\usage{
\method{plot}{gcapthist}(x, which = c("traps", "arrays", "detections"),
  sessions = NULL, ...)

\method{points}{gcapthist}(x, which = c("traps", "arrays", "detections"),
  sessions = NULL, ...)

geom_capthist(x, which = c("traps", "arrays"), sessions = NULL, ...)
}
\arguments{
\item{x}{a \link[gibbonsecr]{gcapthist} object}

\item{which}{type of plot (see Details)}

\item{sessions}{the sessions to plot (default is all sessions)}

\item{...}{additional arguments to pass to \link[graphics]{plot.default},
\link[graphics]{points.default} or \link[ggplot2]{geom_point}}
}
\description{
Plotting methods for gibbonsecr capthist objects.
}
\details{
The following options are available for the \code{which} argument:
  \tabular{ll}{
   \code{traps}      \tab Listening post locations \cr
   \code{arrays}     \tab Centre points of listening post arrays \cr
   \code{detections} \tab Arrows showing the estimated bearings, if present
   (with arrow length equal to to the estimated distance, if present) \cr
  }
}
\examples{
\dontrun{

data(N.annamensis)
capthist = N.annamensis.capthist

# plot methods -----------------------------------------------------------------

# listening posts
plot(capthist)

# arrays
plot(capthist, "arrays")

# overlay
plot(capthist, "arrays", cex = 4, pch = 1)
points(capthist)
points(capthist, sessions = 1:3, col = "red", cex = 1)
points(capthist, "arrays", sessions = 4, col = 4, cex = 4, pch = 19)

# ggplot methods ---------------------------------------------------------------

library(ggplot2)

# listening posts
ggplot() + coord_fixed() +
    geom_capthist(capthist) +
    labs(x = "Longitude", y = "Latitude")

# arrays
ggplot() + coord_fixed() +
    geom_capthist(capthist, "arrays") +
    labs(x = "Longitude", y = "Latitude")

# overlay
ggplot() + coord_fixed() +
    geom_capthist(capthist) +
    geom_capthist(capthist, "arrays", pch = 1, cex = 9) +
    geom_capthist(capthist, sessions = 1:3, col = "red", cex = 1) +
    geom_capthist(capthist, "arrays", sessions = 4, col = 4, cex = 9, pch = 19) +
    labs(x = "Longitude", y = "Latitude")
}


}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link[gibbonsecr]{plot.gmask}
}

