% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.r
\name{plot.gsecr}
\alias{plot.gsecr}
\alias{geom_fit}
\title{Plot a gibbonsecr fitted model}
\usage{
\method{plot}{gsecr}(x, which = c("detfunc", "detsurf", "bearings",
  "distances", "densurf", "locations"), session = 1, ci = FALSE,
  level = 0.95, method = c("delta", "boot"), nboot = 999,
  newdata = NULL, add = FALSE, ...)

geom_fit(x, which = c("detfunc", "detsurf", "bearings", "distances",
  "densurf"), session = 1, ci = FALSE, level = 0.95,
  method = c("delta", "boot"), nboot = 999, contour = FALSE,
  newdata = NULL, ...)
}
\arguments{
\item{x}{a \link[gibbonsecr]{gsecr} object}

\item{which}{plot type}

\item{session}{session to plot (integer or session name)}

\item{ci}{if \code{TRUE}, confidence intervals will be plotted}

\item{level}{confidence level for confidence intervals}

\item{method}{confidence interval estimation method}

\item{nboot}{number of bootstrap samples to use when \code{method = "boot"}}

\item{newdata}{optional dataframe of covariate values}

\item{add}{if \code{TRUE}, adds to existing plot}

\item{...}{additional arguments to be passed to the underlying plotting
method (see Details)}

\item{contour}{if \code{TRUE}, a contour plot will be drawn instead of a
heatmap (for \code{which = "detsurf"} or \code{which = "densurf"})}
}
\description{
Plotting methods for gibbonsecr fitted model objects.
}
\examples{
\dontrun{

data(N.annamensis)
fit = N.annamensis.fit

# plot methods -----------------------------------------------------------------

# > detection function ----

par(mfrow = c(1,1), mar = c(4,4,2,1))
plot(fit)
plot(fit, ci = TRUE)
plot(fit, ci = TRUE, method = "boot")

# > bearing error ----

plot(fit, "bearings")
plot(fit, "bearings", ci = TRUE)
plot(fit, "bearings", ci = TRUE, method = "boot")

# > detection surface ----

plot(fit, "detsurf", session = 1)
points(fit$capthist, session = 1)

# > density surface ----

plot(fit, "densurf")

# ggplot methods ---------------------------------------------------------------

# > detection function ----

ggplot() +
    geom_fit(fit) +
    labs(x = "Radial distance (m)", y = "Detection probability")

ggplot() +
    geom_fit(fit) +
    geom_fit(fit, ci = TRUE) +
    labs(x = "Radial distance (m)", y = "Detection probability")

ggplot() +
    geom_fit(fit) +
    geom_fit(fit, ci = TRUE, method = "boot") +
    labs(x = "Radial distance (m)", y = "Detection probability")

# > bearing error ----

ggplot() +
    geom_fit(fit, "bearings") +
    labs(x = "Bearing error (degrees)", y = "Probability density")

ggplot() +
    geom_fit(fit, "bearings") +
    geom_fit(fit, "bearings", ci = TRUE) +
    labs(x = "Bearing error (degrees)", y = "Probability density")

ggplot() +
    geom_fit(fit, "bearings") +
    geom_fit(fit, "bearings", ci = TRUE, method = "boot") +
    labs(x = "Radial distance (m)", y = "Detection probability")

# > detection surface ----

ggplot() + coord_fixed() +
    geom_fit(fit, "detsurf", session = 1) +
    geom_fit(fit, "detsurf", session = 1, contour = TRUE) +
    geom_capthist(fit$capthist, session = 1) +
    scale_fill_distiller(palette = "Spectral", limits = c(0, 1)) +
    labs(x = "Longitude", y = "Latitude")

# > density surface ----

ggplot() + coord_fixed() +
    geom_fit(fit, "densurf") +
    geom_capthist(fit$capthist, "array") +
    scale_fill_distiller(palette = "Spectral") +
    labs(x = "Longitude", y = "Latitude")
}


}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
