% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot_mask}
\alias{plot_mask}
\title{TODO}
\usage{
plot_mask(mask, traps = NULL, region = NULL, add = FALSE,
  covariate = NULL, col = NULL, pch = 15, bty = "n", axes = FALSE,
  xlab = "", ylab = "", main = "", legend = TRUE, xaxs = "r",
  yaxs = "r")
}
\arguments{
\item{mask}{TODO}

\item{traps}{TODO}

\item{region}{TODO}

\item{add}{TODO}

\item{covariate}{TODO}

\item{col}{TODO}

\item{pch}{TODO}

\item{bty}{TODO}

\item{axes}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{main}{TODO}

\item{legend}{TODO}

\item{xaxs}{TODO}

\item{yaxs}{TODO}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

