% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot_shp}
\alias{plot_shp}
\title{TODO}
\usage{
plot_shp(shp, traps = NULL, add = FALSE)
}
\arguments{
\item{shp}{TODO}

\item{traps}{TODO}

\item{add}{TODO}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

