% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot_traps}
\alias{plot_traps}
\title{TODO}
\usage{
plot_traps(x, add = FALSE, pch = 15, cex = 0.5, col = 1, ...)
}
\arguments{
\item{x}{TODO}

\item{add}{TODO}

\item{pch}{TODO}

\item{cex}{TODO}

\item{col}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

