% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{simulate.gibbonsecr_fit}
\alias{simulate.gibbonsecr_fit}
\title{Simulate data from a fitted model}
\usage{
\method{simulate}{gibbonsecr_fit}(object, nsim = 1, seed = NULL,
  buffer = 6000, spacing = 10, beta = NULL, debug = FALSE, ...)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned
from the \code{\link{gibbonsecr_fit}} function)}

\item{nsim}{number of data sets to simulate (currently fixed at 1)}

\item{seed}{an integer to initialize the random number generator}

\item{buffer}{buffer distance for use in generating the underlying population
(see Details)}

\item{spacing}{TODO}

\item{beta}{TODO}

\item{debug}{TODO}

\item{...}{additional arguments (not used)}
}
\description{
Simulate capture history data from a model fitted with
  \code{\link{gibbonsecr_fit}}
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

