% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{subset_capthist}
\alias{subset_capthist}
\title{SUbset a multi-session capthist}
\usage{
subset_capthist(capthist, index)
}
\arguments{
\item{capthist}{a \code{\link{capthist}} object}

\item{index}{logical, integer or character vector indicating which capthist sessions (multi-session) or rows (single-session) to keep}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

