% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{subset_mask}
\alias{subset_mask}
\title{Subset a mask}
\usage{
subset_mask(mask, index)
}
\arguments{
\item{mask}{a \code{\link{mask}} object}

\item{index}{logical, integer or character vector indicating which mask sessions (multi-session) or points (single-session) to keep}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

