% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{summary.gibbonsecr_fit}
\alias{summary.gibbonsecr_fit}
\title{Summarise a fitted model}
\usage{
\method{summary}{gibbonsecr_fit}(object, ...)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned
from the \code{\link{gibbonsecr_fit}} function)}

\item{...}{additional arguments (not used)}
}
\description{
Summarise the results of a model fitted with
  \code{\link{gibbonsecr_fit}}.
}
\examples{
## Not run: 

data(N.annamensis)

fit = gibbonsecr_fit(N.annamensis)

AIC(fit)
coef(fit)
#confint(fit)
#plot(fit)
#predict(fit)
#print(fit)
vcov(fit)

## End(Not run)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

