% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{summary_capthist}
\alias{summary_capthist}
\title{TODO}
\usage{
summary_capthist(capthist)
}
\arguments{
\item{capthist}{a \code{\link{capthist}} object}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

