% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{vcov.gibbonsecr_fit}
\alias{vcov.gibbonsecr_fit}
\title{Compute the variance-covariance matrix}
\usage{
\method{vcov}{gibbonsecr_fit}(object, ...)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned from the \code{\link{gibbonsecr_fit}} function)}

\item{...}{additional arguments (not used)}
}
\description{
Compute the variance-covariance from a model fitted with \code{\link{gibbonsecr_fit}}.
}
\examples{
## Not run: 

data(N.annamensis)

fit = gibbonsecr_fit(N.annamensis)

AIC(fit)
coef(fit)
#confint(fit)
#plot(fit)
#predict(fit)
#print(fit)
vcov(fit)

## End(Not run)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

