% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{AIC.gibbonsecr_fit}
\alias{AIC.gibbonsecr_fit}
\title{Compute the AIC score}
\usage{
\method{AIC}{gibbonsecr_fit}(object, ..., k = 2)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned
from the \code{\link{gibbonsecr_fit}} function)}

\item{...}{additional arguments (not used)}

\item{k}{the penalty per parameter to be used (defaults to k = 2)}
}
\description{
Compute the AIC score for a model fitted with
  \code{\link{gibbonsecr_fit}}.
}
\examples{
## Not run:

library(gibbonsecr)

data(N.annamensis)

fit = gibbonsecr_fit(N.annamensis)

AIC(fit)
coef(fit)
#confint(fit)
#plot(fit)
#predict(fit)
#print(fit)
vcov(fit)

## End(Not run)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

