% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{confint.gibbonsecr_fit}
\alias{confint.gibbonsecr_fit}
\title{Compute confidence intervals for model parameters}
\usage{
\method{confint}{gibbonsecr_fit}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned
from the \code{\link{gibbonsecr_fit}} function)}

\item{parm}{parameters (see documentation for the \code{\link{confint}}
generic)}

\item{level}{the confidence level}

\item{...}{additional arguments (not used)}
}
\description{
Compute confidence intervals stimated parameters from a model
  fitted with \code{\link{gibbonsecr_fit}}.
}
\details{
Uses the delta method to obtain interval estimates on the link
  scale.
}
\examples{
## Not run: 

data(N.annamensis)

fit = gibbonsecr_fit(N.annamensis)

AIC(fit)
coef(fit)
#confint(fit)
#plot(fit)
#predict(fit)
#print(fit)
vcov(fit)

## End(Not run)
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

