% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-package.r
\docType{package}
\name{gibbonsecr-package}
\alias{gibbonsecr-package}
\title{SECR for acoustic gibbons surveys}
\description{
An implementation of Spatially Explicit Capture-Recapture (SECR) for the analyis of data from acoustic population surveys of gibbons.
}
\details{
Describe one and multi-occasion surveys.
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\references{
Some references
}

