% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-import.R
\name{import_data}
\alias{import_data}
\title{Import survey data}
\usage{
import_data(detections, posts, covariates = NULL, details = list())
}
\arguments{
\item{detections}{recapture data and (optionally) estimated bearings and
distances to detected groups}

\item{posts}{listening post coordinates (in metric units)}

\item{covariates}{covariate values (optional)}

\item{details}{list containing the type and units of the estimated bearings
and distances (see Details)}
}
\description{
The main arguments to this function are \code{detections},
  \code{posts} and (optionally) \code{covariates}. The values of these
  arguments can be either (i) a string giving the the file path to a
  \code{.csv} file, or (ii) a data frame (see Details for a full description
  of the required format in each case).

  The function returns a \link[secr]{capthist} object that can be used
  as an input to the \link[gibbonsecr]{gfit} fitting function. The
  capthist object returned differs from the standard format used by the
  \pkg{secr} package in that arrays of estimated bearings and distances are
  attached as attributes (if present in the detections data).
}
\details{
The format of the \code{.csv} files (or data frames) and the
  \code{details} argument are described below:
\describe{
 \item{\strong{detections}}{This should contain one row per detection. For
 example, if group X was detected at array 1 on occasion 1 at posts A and B,
 and on occasion 2 at post C, then this would represent three detections.
 This data should contain the folowing columns:
  \tabular{ll}{
   \code{array}    \tab array ID \cr
   \code{occasion} \tab sampling occasion (must be an integer) \cr
   \code{post}     \tab listening post ID \cr
   \code{group}    \tab group ID \cr
   \code{bearing}  \tab estimated bearing (optional) \cr
   \code{distance} \tab estimated distance (optional) \cr
   \tab\cr\tab\cr
  }
 }
 \item{\strong{posts}}{This should contain one row per listening post and
 contain the folowing columns:
  \tabular{ll}{
   \code{array} \tab array ID \cr
   \code{post}  \tab listening post ID \cr
   \code{x}     \tab listening post x-coordinate (in metric units) \cr
   \code{y}     \tab listening post y-coordinate (in metric units) \cr
   \code{usage} \tab listening post usage (must be a string of 1s and 0s) \cr
  }
  The values in the usage column should reflect the number of occasions. For
  single-occasion surveys all entries in the usage column should be
  \code{"1"}. For multi-occasion surveys the length of each entry in the
  usage column should be equal to the number of occasions for that array. For
  example, in a three-occasion survey if post X was used on all occasions
  then the entry in the usage column should be \code{"111"}, but if it was
  used on occasions 1 and 3 but not on occasion 2 then the entry would be
  \code{"101"}.
 }
 \item{\strong{covariates}}{This should contain one row per listening
 post-occasion combination. For example, if there are 3 listening posts at
 array X and 2 sampling occasions, the covariates data should contain 6 rows
 for that array (irrespective of the number of detections). This data should
 contain the folowing columns:
  \tabular{ll}{
   \code{array}        \tab the array ID \cr
   \code{occasion}     \tab the sampling occasion (must be an integer) \cr
   \code{post}         \tab the listening post ID \cr
   \code{<covariate1>} \tab values for covariate 1 \cr
   \code{<covariate1>} \tab values for covariate 2 \cr
   \code{...}          \tab ... \cr
  }
  For advice on model specification using covariates see
  \link[gibbonsecr]{gibbonsecr-covariates}.
 }
 \item{\strong{details}}{ The value of the \code{details} argument should be
 a list. If there are no estimated bearings and no estimated distances then
 this should be an empty list (the default).

 If the detections data contains estimated bearings, then one element of the
 \code{details} list should be named \code{"bearings"} and the value of this
 element should be a list with element names \code{"units"} and
 \code{"type"}. The \code{"units"} element can be set to either
 \code{"degrees"} (the default) or \code{"radians"}, and the \code{"type"}
 element should be set to \code{"continuous"} (currently the only available
 option). If bearings data exist and the \code{"bearings"} element is
 \code{NULL} then the default values are assumed.

 If the detections data contains estimated bearings, then one element of the
 \code{details} list should be named \code{"distances"} and the value of this
 element should be a list with element names \code{"units"} and
 \code{"type"}. The \code{"units"} element can be set to either \code{"km"}
 (the default) or \code{"m"}, and the \code{"type"} element should be set to
 \code{"continuous"} (currently the only available option). If distances data
 exist and the \code{"distances"} element is \code{NULL} then the default
 values are assumed.
 }
}
}
\examples{
\dontrun{

# import from csv files --------------------------------------------------------

folder = system.file("extdata/N.annamensis", package = "gibbonsecr")
capthist = import_data(
    detections = file.path(folder, "detections.csv"),
    posts      = file.path(folder, "posts.csv"),
    covariates = file.path(folder, "covariates.csv"),
    details    = list(bearings = list(units = "degrees",
                                      type = "continuous"))
)
summary(capthist)

# import capthist data frames --------------------------------------------------

detections = read.csv(file.path(folder, "detections.csv"))
posts      = read.csv(file.path(folder, "posts.csv"))
covariates = read.csv(file.path(folder, "covariates.csv"))
head(detections)
head(posts)
head(covariates)
# using default details
capthist = import_data(
    detections = detections,
    posts      = posts,
    covariates = covariates,
    details    = list(bearings = list(units = "degrees",
                                      type = "continuous"))
)
summary(capthist)
}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link[gibbonsecr]{plot.gcapthist}, \link[gibbonsecr]{import_shp}
}

