% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_main.r
\name{import_shp}
\alias{import_shp}
\title{Import shp files}
\usage{
import_shp(path, verbose = FALSE, ...)
}
\arguments{
\item{path}{file path}

\item{verbose}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

