% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-import.R
\name{import_shp}
\alias{import_shp}
\title{Import ESRI shapefiles}
\usage{
import_shp(x, fortify = TRUE, ...)
}
\arguments{
\item{x}{file path of the shapefile including the file name (the file
extension is not required since it is assumed to be \code{.shp}), or a
\link[sp]{SpatialPolygons} or \link[sp]{SpatialPoints} object}

\item{fortify}{if \code{TRUE}, a 'fortified' version of the spatial object
will be returned (see Details)}

\item{...}{additional arguments to pass to \link[raster]{shapefile}}
}
\description{
This is a wrapper for the \link[raster]{shapefile} function
  (from the \pkg{raster} package). It performs additonal checks on data
  covariates (converting all logical and character variables to factors) and
  prepares a 'fortified' version of \link[sp]{SpatialPolygons} objects data
  to allow \pkg{ggplot}-style plotting.
}
\details{
A list with two elements:
\tabular{ll}{
  \code{sp} \tab an object of class \link[sp]{SpatialPolygons} or
  \link[sp]{SpatialPoints} \cr
  \code{fortified} \tab for polygon shapefiles, this will be the result of
  applying the \link[ggplot2]{fortify} function to the \code{sp} element \cr
}
}
\examples{
\dontrun{

folder = system.file("extdata/N.annamensis", package = "gibbonsecr")
region = import_shp(file.path(folder, "region.shp"))
habitat = import_shp(file.path(folder, "habitat.shp"))
}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link[gibbonsecr]{plot.gshp}, \link[gibbonsecr]{import_data}
}

