% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-data.r
\docType{data}
\name{N.annamensis}
\alias{N.annamensis}
\title{Example survey data}
\format{A object of class \code{\link{capthist}}.}
\source{
Data were provided by Dr Ben Rawson, IUCN SSC Primate Specialist
  Group, Southeast Asia/Indochina.
}
\usage{
data(N.annamensis)
}
\description{
Data from 2010 survey of \emph{Nomascus anamensis}, Veun Sai Reserve, NW Cambodia.
}
\references{
Thinh, V. N., Mootnick, A. R., Thanh, V. N., and Roos, C. 2010.
  A new species of crested gibbon, from the central Annamite mountain range.
  \emph{Vietnamese Journal of Primatology}, 1(4):112.
}

