% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_plots.r
\name{plot.gibbonsecr_fit}
\alias{plot.gibbonsecr_fit}
\title{Plot a fitted model}
\usage{
\method{plot}{gibbonsecr_fit}(x, which = c("detectfn", "pdot", "bearings",
  "distances", "density"), session = 1, newdata = NULL,
  use.global.par.settings = TRUE, ...)
}
\arguments{
\item{x}{a \code{gibbonsecr} model object (i.e. an object returned from the
\code{\link{gibbonsecr_fit}} function)}

\item{which}{TODO}

\item{session}{TODO}

\item{use.global.par.settings}{TODO}

\item{...}{additional arguments (not used)}
}
\description{
TODO
}
\details{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

