% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.r
\name{plot.gmask}
\alias{plot.gmask}
\alias{geom_mask}
\title{Plot a gibbonsecr mask}
\usage{
\method{plot}{gmask}(x, covariate = NULL, sessions = NULL, ...)

geom_mask(x, covariate = NULL, sessions = NULL, contour = FALSE, ...)
}
\arguments{
\item{x}{a \link[gibbonsecr]{gmask} object}

\item{covariate}{name of covariate to use for colouring}

\item{sessions}{sessions to plot (as integer indices, logical indices or
session names)}

\item{...}{additional arguments to pass to \link[secr]{plot.mask} or
\link[ggplot2]{geom_point}}

\item{contour}{if \code{TRUE}, contour lines will be plotted}
}
\description{
Plotting functions for gibbonsecr mask objects. The plot method
is a wrapper for \link[secr]{plot.mask} and geom_mask is a wrapper for
\link[ggplot2]{geom_point} (or \link[ggplot2]{geom_contour} when
\code{contour = TRUE})
}
\examples{
\dontrun{

data(N.annamensis)
mask = N.annamensis.mask
capthist = N.annamensis.capthist

# plot methods -----------------------------------------------------------------

par(mar = c(2,2,1,1))

plot(mask)
points(capthist, cex = 0.5)

plot(mask, covariate = "x", legend = FALSE)
points(capthist, cex = 0.5)

plot(mask, covariate = "y", legend = FALSE)
points(capthist, cex = 0.5)

plot(mask, covariate = "habitat", legend = FALSE)
points(capthist, cex = 0.5)

# ggplot methods ---------------------------------------------------------------

library(ggplot2)

ggplot() + coord_fixed() +
    geom_mask(mask) +
    geom_capthist(capthist) +
    labs(x = "Longitude", y = "Latitude")

ggplot() + coord_fixed() +
    geom_mask(mask, covariate = "x") +
    geom_capthist(capthist) +
    scale_fill_distiller(palette = "Spectral") +
    labs(x = "Longitude", y = "Latitude", fill = "Longitude")

ggplot() + coord_fixed() +
    geom_mask(mask, covariate = "y") +
    geom_capthist(capthist) +
    scale_fill_distiller(palette = "Spectral") +
    labs(x = "Longitude", y = "Latitude", fill = "Latitutde")

ggplot() + coord_fixed() +
    geom_mask(mask, covariate = "habitat") +
    geom_capthist(capthist) +
    scale_fill_manual(values = c("darkgreen", "lightgreen")) +
    labs(x = "Longitude", y = "Latitude", fill = "habitat")
}


}
\seealso{
\link[gibbonsecr]{plot.gcapthist}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
