% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.r
\name{plot.gshp}
\alias{geom_shp}
\alias{plot.gshp}
\title{Plot an imported shapfile}
\usage{
\method{plot}{gshp}(x, covariate = NULL, contour = FALSE, ...)

geom_shp(x, covariate = NULL, contour = FALSE, ...)
}
\arguments{
\item{x}{a \link[gibbonsecr]{gshp} object}

\item{covariate}{name of covariate to use for colouring}

\item{contour}{if \code{TRUE}, contour lines will be plotted}

\item{...}{additional arguments to be passed to the underlying plotting
method (see Details)}
}
\description{
Plotting methods for imported shapefiles.
}
\examples{
\dontrun{

data(N.annamensis)

# plot methods -----------------------------------------------------------------

plot(N.annamensis.region, col = "grey")

plot(N.annamensis.habitat, covariate = "habitat", col = c("darkgreen", "lightgreen"))
plot(N.annamensis.region, add = TRUE)

# ggplot methods ---------------------------------------------------------------

library(ggplot2)

ggplot() + coord_fixed() +
    geom_shp(N.annamensis.region)

ggplot() + coord_fixed() +
    geom_shp(N.annamensis.habitat, covariate = "habitat") +
    scale_fill_manual(values = c("darkgreen", "lightgreen"))

ggplot() + coord_fixed() +
    geom_shp(N.annamensis.habitat, covariate = "habitat") +
    geom_shp(N.annamensis.region) +
    scale_fill_manual(values = c("darkgreen", "lightgreen"))
}
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}

