% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gibbonsecr_1.0-functions_methods.r
\name{simulate.gibbonsecr_fit}
\alias{simulate.gibbonsecr_fit}
\title{TODO}
\usage{
\method{simulate}{gibbonsecr_fit}(object, nsim = 1, seed = NULL,
  buffer = 6000, spacing = 10, beta = NULL, ...)
}
\arguments{
\item{object}{a \code{gibbonsecr_fit} model object (i.e. an object returned from the \code{\link{gibbonsecr_fit}} function)}

\item{nsim}{TODO}

\item{seed}{TODO}

\item{buffer}{TODO}

\item{spacing}{TODO}

\item{beta}{TODO}

\item{...}{additional arguments (not used)}
}
\description{
TODO
}
\author{
Darren Kidney \email{darrenkidney@googlemail.com}
}
\seealso{
\link{gibbonsecr_fit}
}

